+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.1
+
+- Add support for custom headers
+- Fix a bug which won't let you download small files
+
 ## 1.0.0
 
 - Initial release after rename from `m-downloader`